refactor: separate code to helper
